Updated for our belief system
We decided "Open Source" and "Copyright" are an oxymoron.  Feel free to fork, edit, or even contribute to our project in any way you feel fit.  We won't come after you with pitchforks and flames.

Also, I added a real description of what Tentacles are to us and our users.